+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/task02/ArrayIndexOutOfBoundsExceptionTest.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.exception_handling;
+package com.rakovets.course.java.core.practice.exception_handling.task02;
 
 public class ArrayIndexOutOfBoundsExceptionTest {
     public void firstMethod() {
