+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/task03/OwnClassHeirException.java
@@ -1,0 +1,7 @@
+package com.rakovets.course.java.core.practice.exception_handling.task03;
+
+public class OwnClassHeirException extends Exception {
+    public OwnClassHeirException(String message) {
+        super(message);
+    }
+}
